@@ -19,55 +19,56 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  assorted_layout_widgets: ^2.0.0
-  bloc: ^6.1.0
-  cached_network_image: ^2.2.0+1
+  assorted_layout_widgets: ^3.0.0
+  bloc: ^6.1.1
+  cached_network_image: ^2.5.0
   calendar_strip: ^1.0.7
   carousel_pro: ^1.0.0
   checkbox_formfield: ^0.1.0+3
-  cloud_firestore: ^0.13.7
-  country_code_picker: ^1.3.12+1
-  cupertino_icons: ^0.1.2
+  cloud_firestore: ^0.14.4
+  country_code_picker: ^1.7.0
+  cupertino_icons: ^1.0.0
   equatable: ^1.2.5
-  firebase_analytics: ^5.0.14
-  firebase_auth: ^0.16.1
-  firebase_core: ^0.4.4
-  firebase_messaging: ^6.0.1
-  firebase_remote_config: ^0.3.1
+  firebase_analytics: ^6.3.0
+  firebase_auth: ^0.18.4+1
+  firebase_core: ^0.5.3
+  firebase_messaging: ^7.0.3
+  firebase_remote_config: ^0.4.3
   flare_flutter: ^2.0.6
-  flutter_bloc: ^6.1.0
-  flutter_layout_grid: ^0.10.1
-  flutter_rating_bar: ^3.0.1+1
+  flutter_bloc: ^6.1.1
+  flutter_layout_grid: ^0.10.2
+  flutter_rating_bar: ^3.2.0+1
   flutter_svg: ^0.19.1
-  font_awesome_flutter: ^8.8.1
+  font_awesome_flutter: ^8.11.0
   geocoding: ^1.0.5
   geolocator: ^6.1.13
-  google_fonts: ^1.1.0
-  hive: ^1.4.2
+  google_fonts: ^1.1.1
+  hive: ^1.4.4+1
   hive_flutter: ^0.3.1
-  http: ^0.12.1
+  http: ^0.12.2
   icon_shadow: ^1.0.1
   intl: ^0.16.1
-  intro_slider: ^2.3.1
+  intro_slider: ^2.3.3
   keyboard_avoider: ^0.1.2
-  logger: ^0.9.1
-  maps_launcher: ^1.2.2+1
-  marquee_widget: ^0.8.0
-  ordered_set: ^2.0.0
-  path_drawing: ^0.4.1
-  pin_code_fields: ^5.0.1
-  provider: ^4.1.1
-  shimmer: ^1.1.1
+  logger: ^0.9.4
+  maps_launcher: ^1.2.2+2
+  marquee_widget: ^1.0.0
+  ordered_set: ^2.0.1
+  path_drawing: ^0.4.1+1
+  pin_code_fields: ^6.0.2
+  provider: ^4.3.2+3
+  shimmer: ^1.1.2
   url_launcher: ^5.5.0
   webview_flutter: ^1.0.7
+  flutter_map: ^0.10.1+1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   bloc_test: ^7.1.0
-  build_runner: ^1.10.1
-  hive_generator: ^0.7.1
-  test: ^1.15.2
+  build_runner: ^1.10.11
+  hive_generator: ^0.8.2
+  test: ^1.15.7
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
